requirements.txt
# GCP Dependencies
google-cloud-pubsub==2.18.4
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0

# Apache Beam and Dataflow
apache-beam[gcp]==2.53.0

# Data Processing and ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Optional but recommended packages
python-dateutil==2.8.2
pytz==2023.3.post1
grpcio==1.59.3
protobuf==4.25.1

# Development and testing
pytest==7.4.3
black==23.11.0
pylint==3.0.2